+++ imports/api/analytics/analytics.js
@@ -1,9 +1,9 @@
 import { Mongo } from 'meteor/mongo';
 import { Blockscon } from '../blocks/blocks.js';
 
-export const Sales = new Mongo.Collection('sales');
+export const Analytics = new Mongo.Collection('nft-analytics');
 
-Sales.helpers({
+Analytics.helpers({
     block(){
         return Blockscon.findOne({height:this.height});
     }
