--- imports/api/listings/server/publications.js
@@ -1,9 +1,0 @@
-import { Listings } from '../listings.js';
-
-publishComposite('Listings.list', function(limit = 30){
-    return {
-        find(){
-            return Listings.find({}, { sort: { ID: 1 } });
-        }
-    }
-});
