--- imports/api/sales/server/publications.js
@@ -1,9 +1,0 @@
-import { Sales } from '../sales.js';
-
-publishComposite('Sales.list', function(limit = 30){
-    return {
-        find(){
-            return Sales.find({}, { sort: { ID: 1 } });
-        }
-    }
-});
