+++ imports/startup/server/register-api.js
@@ -40,8 +40,5 @@ import '../../api/accounts/server/methods.js';
 
 import '../../api/coin-stats/server/methods.js';
 
-import '../../api/listings/server/methods.js';
-import '../../api/listings/server/publications.js';
-
-import '../../api/sales/server/methods.js';
-import '../../api/sales/server/publications.js';
+import '../../api/analytics/server/methods.js';
+import '../../api/analytics/server/publications.js';
